Fixes: 1) Problem with checksum for Quickpay Form, 2). Added correct Post template to Updates folder
