+++ src/Hash.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
 
@@ -6,7 +8,29 @@ namespace Dynamicweb.Ecommerce.CheckoutHandlers.QuickPayPaymentWindow;
 
 internal static class Hash
 {
-    public static string ByteArrayToHexString(byte[] bytes)
+    public static string ComputeHash(string key, IDictionary<string, string> formValues)
+    {
+        string message = GetMacString(formValues);
+
+        return ComputeHash(key, message);
+    }
+
+
+    public static string ComputeHash(string key, string message)
+    {
+        var encoding = new UTF8Encoding();
+        byte[] byteKey = encoding.GetBytes(key);
+
+        using (HMACSHA256 hmac = new HMACSHA256(byteKey))
+        {
+            var messageBytes = encoding.GetBytes(message);
+            var hashedBytes = hmac.ComputeHash(messageBytes);
+
+            return ByteArrayToHexString(hashedBytes);
+        }
+    }
+
+    private static string ByteArrayToHexString(byte[] bytes)
     {
         var result = new StringBuilder();
         foreach (byte b in bytes)
@@ -17,29 +41,25 @@ internal static class Hash
         return result.ToString();
     }
 
-    public static string ComputeHash(string key, Stream message)
+    private static string GetMacString(IDictionary<string, string> formValues)
     {
-        var encoding = new UTF8Encoding();
-        var byteKey = encoding.GetBytes(key);
+        var excludeList = new List<string> { "MAC" };
+        var keysSorted = formValues.Keys.ToArray();
+        Array.Sort(keysSorted, StringComparer.Ordinal);
 
-        using (HMACSHA256 hmac = new HMACSHA256(byteKey))
+        var message = new StringBuilder();
+        foreach (string key in keysSorted)
         {
-            var hashedBytes = hmac.ComputeHash(message);
-            return ByteArrayToHexString(hashedBytes);
+            if (excludeList.Contains(key))
+                continue;
+
+            if (message.Length > 0)
+                message.Append(" ");
+
+            var value = formValues[key];
+            message.Append(value);
         }
-    }
 
-    public static string ComputeHash(string key, string message)
-    {
-        var encoding = new UTF8Encoding();
-        var byteKey = encoding.GetBytes(key);
-
-        using (HMACSHA256 hmac = new HMACSHA256(byteKey))
-        {
-            var messageBytes = encoding.GetBytes(message);
-            var hashedBytes = hmac.ComputeHash(messageBytes);
-
-            return ByteArrayToHexString(hashedBytes);
-        }
+        return message.ToString();
     }
 }
