--- src/Updates/Post.html
@@ -1,20 +1,0 @@
-﻿<form action="https://payment.quickpay.net" method="post">
-    <input type="hidden" name="version" value="v10">
-    <input type="hidden" name="agreement_id" value="<!--@QuickPayPaymentWindow.agreement_id-->" />
-    <input type="hidden" name="merchant_id" value="<!--@QuickPayPaymentWindow.merchant_id-->" />
-    <input type="hidden" name="language" value="<!--@QuickPayPaymentWindow.language-->" />
-    <input type="hidden" name="order_id" value="<!--@QuickPayPaymentWindow.order_id-->" />
-    <input type="hidden" name="amount" value="<!--@QuickPayPaymentWindow.amount-->" />
-    <input type="hidden" name="currency" value="<!--@QuickPayPaymentWindow.currency-->" />
-    <input type="hidden" name="continueurl" value="<!--@QuickPayPaymentWindow.continueurl-->" />
-    <input type="hidden" name="cancelurl" value="<!--@QuickPayPaymentWindow.cancelurl-->" />
-    <input type="hidden" name="callbackurl" value="<!--@QuickPayPaymentWindow.callbackurl-->" />
-    <input type="hidden" name="autocapture" value="<!--@QuickPayPaymentWindow.autocapture-->" />
-    <input type="hidden" name="autofee" value="<!--@QuickPayPaymentWindow.autofee-->" />
-    <input type="hidden" name="payment_methods" value="<!--@QuickPayPaymentWindow.payment_methods-->" />
-    <input type="hidden" name="branding_id" value="<!--@QuickPayPaymentWindow.branding_id-->" />
-    <input type="hidden" name="google_analytics_tracking_id" value="<!--@QuickPayPaymentWindow.google_analytics_tracking_id-->" />
-    <input type="hidden" name="google_analytics_client_id" value="<!--@QuickPayPaymentWindow.google_analytics_client_id-->" />
-    <input type="hidden" name="checksum" value="<!--@QuickPayPaymentWindow.checksum-->" />
-    <input type="submit" value="Start payment" />
-</form>
